fix: BaseEntity의 id 타입 수정으로 인한 오류 수정

+++ back/src/main/java/com/back/domain/member/member/entity/Member.java
@@ -35,7 +35,7 @@ public class Member extends BaseEntity {
     @Enumerated(EnumType.STRING)
     private Role role = Role.MEMBER;
 
-    public Member(int id, String email, String nickname) {
+    public Member(long id, String email, String nickname) {
         setId(id);
         this.email = email;
         this.name = nickname;
